add singularity support, generalize code to containers, simulate error and add test,
