+++ src/cpac/api/backends/__init__.py
@@ -1,0 +1,7 @@
+from .docker import DockerSchedule
+from .singularity import SingularityBackend
+
+available_backends = {
+    'docker': DockerSchedule,
+    'singularity': SingularityBackend,
+}
