+++ src/cpac/api/backends/utils.py
@@ -1,7 +1,16 @@
 import asyncio
 import logging
+import socket
+from contextlib import closing
 
 logger = logging.getLogger(__name__)
+
+
+def find_free_port():
+    with closing(socket.socket(socket.AF_INET, socket.SOCK_STREAM)) as s:
+        s.bind(('', 0))
+        s.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
+        return s.getsockname()[1]
 
 
 async def consume(async_gen):
